added very basic initial layout for level 3
Reminder to future me: In order to get through the exit, the player must have 2 keys. This means that the key within the locked room is required in order to progress. If number of keys is 0, tell the player that they require a keycard. If number of keys = 1, tell the player they have invalid keycard. If number of keys = 2, the exit will become usable.
When the player dies on this level, set number of keys to 0.
If they die on the next level, set number of keys to 2 (to allow backtracking)